+++ src/App.vue
@@ -1,7 +1,10 @@
 <script setup lang="ts">
+import Menu from './components/Menu.vue';
+
 
 </script>
 
 <template>
+  <Menu/>
   <RouterView/>
 </template>
